Rename surface mesh task to surface_mesh_multiscale.
Update imports and tests accordingly.
